server/requirements.txt
openai
flask
flask-cors
pydub
flask[async]
flask-sqlalchemy
flask_migrate
flask_login
flask_bcrypt
oauthlib
requests
PyOpenSSL
pyjwt
google-auth-oauthlib
google-auth